pyproject.toml
[project]
name = "chuni-parsing-library"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.1.1",
    "wannacri>=0.3.1",
]
